+++ CHANGELOG.md
@@ -9,9 +9,13 @@
 [GitHub commits](https://github.com/netket/netket/compare/v3.0...master).
 
 ### New features
-
+* Added Conversion methods `to_qobj()` to operators and variational states, that produce QuTiP's qobjects.
+* A function `nk.nn.activation.reim` has been added that transforms a nonlinearity to act seperately on the real and imaginary parts
+* Nonlinearities `reim_selu` and `reim_relu` have been added
 
 ### Breaking Changes
+* The default initializer for `netket.models.GCNN` has been changed to from `jax.nn.selu` to `netket.nn.reim_selu` [#892](https://github.com/netket/netket/pull/892)
+* `netket.nn.initializers` has been deprecated in favor of `jax.nn.initializers` [#935](https://github.com/netket/netket/pull/935).
 
 ### Internal Changes